Move decoder block stage into new file
