+++ models/demos/deepseek_v3_b1/demo/pipeline.py
@@ -4,7 +4,7 @@
 
 """
 Pipeline orchestration: configuration, factory functions, Pipeline.
-Stage kinds (Embedding, LMHead, Passthrough) live in stage.py.
+Stage kinds (Embedding, LMHead, Passthrough) live in stage.py; decoder/dense blocks in decoder_stage.py.
 """
 
 from __future__ import annotations
@@ -14,6 +14,7 @@ from typing import Any, Callable
 from loguru import logger
 
 import ttnn
+from models.demos.deepseek_v3_b1.demo.decoder_stage import DecoderBlockStage, DenseBlockStage
 from models.demos.deepseek_v3_b1.demo.stage import (
     EmbeddingStage,
     LMHeadStage,
@@ -24,7 +25,6 @@ from models.demos.deepseek_v3_b1.demo.stage import (
 )
 from models.demos.deepseek_v3_b1.demo.weight_provider import WeightProvider
 from models.demos.deepseek_v3_b1.micro_ops.pipeline_block.op import PipelineBlock
-from models.demos.deepseek_v3_b1.tests.unit_tests.test_decoder_block_api import DecoderBlockStage, DenseBlockStage
 
 
 def create_fabric_router_config(max_payload_size: int) -> Any:
